fix: CI mergeability — lint limits and graph editor draft assertion
Split prompt-syntax responsive script tests to satisfy noExcessiveLinesPerFile,
add save-hook test suppression, and assert toolbar omits draft actions after save
instead of removed "No draft changes" copy.
