+++ ui/src/features/workflow-activity/components/react-flow-current-activity-card-edit-integration.test.tsx
@@ -524,9 +524,14 @@ describe("ReactFlowCurrentActivityCard edit integration", () => {
         screen.queryByRole("button", { name: "worker:spare" }),
       ).toBeNull();
     });
-    expect(
-      within(toolbar).getByText("No draft changes", { exact: true }),
-    ).toBeTruthy();
+    await waitFor(() => {
+      expect(
+        within(toolbar).queryByRole("button", { name: "Discard changes" }),
+      ).toBeNull();
+      expect(
+        within(toolbar).queryByRole("button", { name: "Save changes" }),
+      ).toBeNull();
+    });
   });
 
   it("opens save confirmation from the activity card host portaled to document.body", async () => {
